don't request shared irq, and make iobase long for 64 bit arches
git-svn-id: https://svn.code.sf.net/p/linux-gpib/code/trunk@1223 b61deed1-901b-0410-a6dd-e94ecbdaea89

+++ linux-gpib/driver/hp82335/hp82335.h
@@ -27,7 +27,7 @@ typedef struct
 {
 	tms9914_private_t tms9914_priv;
 	unsigned int irq;
-	unsigned int raw_iobase;
+	unsigned long raw_iobase;
 } hp82335_private_t;
 
 // interfaces
